edit projection matrix

+++ cs260_4/client/backend/graphics/renderer.h
@@ -43,7 +43,7 @@ class renderer
 	const glm::vec3        _bgcol, _objcol;
 
 	window&   _winref;
-	int       _width, _height;
+	float     _width, _height;
 	shader    _shader;
 	data      _rectdata, _circledata;
 	SHADERLOC _mmat, _vmat, _pmat;
